+++ frontend/src/RouterRoot.vue
@@ -3,6 +3,6 @@
 </script>
 
 <template>
-  <router-view :key="$route.fullPath" />
+  <router-view />
 </template>